fix: move language service to `monaco-graphql`
even better! this entirely removes `picomatch` from the `codemirror-graphql` depdency chain
