+++ packages/graphql-language-service/package.json
@@ -34,8 +34,7 @@
     "graphql-language-service-interface": "^2.10.1",
     "graphql-language-service-parser": "^1.10.4",
     "graphql-language-service-types": "^1.8.7",
-    "graphql-language-service-utils": "^2.7.1",
-    "picomatch-browser": "^2.2.5"
+    "graphql-language-service-utils": "^2.7.1"
   },
   "devDependencies": {
     "@types/picomatch": "^2.3.0",
