+++ packages/graphql-language-service/src/index.ts
@@ -12,14 +12,6 @@
  *
  * TODO: retire `graphql-language-service-{parser,interface,types,utils}` and merge with this workspace
  */
-export type {
-  SchemaConfig,
-  BaseSchemaConfig,
-  SchemaLoader,
-} from './schemaLoader';
-export { defaultSchemaLoader } from './schemaLoader';
-
-export type { GraphQLLanguageConfig } from './LanguageService';
 /**
  * A whole bunch of the key language services
  */
@@ -41,6 +33,7 @@ export {
   GraphQLLanguageService,
   SEVERITY,
   Severity,
+  HoverConfig,
   SeverityEnum,
   DIAGNOSTIC_SEVERITY,
   DefinitionQueryResult,
