+++ packages/monaco-graphql/package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "main": "dist/monaco.contribution.js",
   "module": "esm/monaco.contribution.js",
-  "types": "dist/monaco.contribution.d.ts",
+  "types": "esm/monaco.contribution.d.ts",
   "contributors": [
     {
       "name": "Peng Lyu",
@@ -26,8 +26,9 @@
     "src"
   ],
   "dependencies": {
-    "graphql-language-service": "^4.1.2",
-    "graphql-language-service-utils": "^2.7.1"
+    "graphql-language-service": "^4.1.1",
+    "graphql-language-service-utils": "^2.7.1",
+    "picomatch-browser": "^2.2.5"
   },
   "devDependencies": {
     "graphql": "16.0.0-experimental-stream-defer.5",
