+++ packages/monaco-graphql/src/typings/picomatch-browser.ts
@@ -1,0 +1,1 @@
+declare module 'picomatch-browser';
